Refactored name package opdracht 3 tm 7
