+++ leertaak2/src/main/java/madeexercises/opdracht3tm7/GetalRij.java
@@ -1,4 +1,4 @@
-package madeexercises.opdracht3;
+package madeexercises.opdracht3tm7;
 
 import java.util.*;
