+++ leertaak2/src/test/java/madeexercises/opdracht3tm7/GetalRijTest.java
@@ -1,4 +1,4 @@
-package madeexercises.opdracht3;
+package madeexercises.opdracht3tm7;
 
 import org.junit.Test;
